add more tab names
