+++ frontend/src/app/layout.tsx
@@ -29,7 +29,7 @@ export const metadata: Metadata = {
   title: "Party Registration",
   description: "Party Registration",
   icons: {
-    icon: "/Home1CarolinaBlue.jpg",
+    icon: "/Home1CarolinaBlue.png",
   },
 };
